+++ dwds/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Properly handle `requireJS` errors during hot restarts.
 - Fix an issue where calling add breakpoint in quick succession
   would corrupt the internal state.
+- Now log the encoded URI of the debug service to both the terminal
+  and application console.
 
 ## 7.1.0